Add resolution for arrays.
Note that by default this will not resolve arrays of structs: requests for int[]
or char[] or DateTime[] will return null. We need to do it this way because
System.String has a constructor of type char[], and resolving this would mislead
FactoryFactory into thinking that it can resolve strings when it can't. This
would fail Microsoft's tests.